enhance(server): onlyServer / onlyQueue オプションを復活
